requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
phonenumbers>=9.0.26
email-validator>=2.3.0
